+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Version 1.0.0 (Not yet Released)
 
 * Reduce Memory Footprint of VnodeRepairState - Issue #1586
+* Avoid Repeated Gauge Registration in TableRepairMetricsImpl - Issue #1584
 * Pre-size LinkedHashMap Allocation in VnodeRepairStatesImpl.Builder - Issue #1587
 * Optimize VnodeRepairStatesImpl combineWithRepairedAt Allocations - Issue #1582
 * Implement ReplicaSet Interning for VnodeRepairState Instances - Issue #1581